+++ test/kubernetes/e2e/features/aiextension/testdata/routes-basic.yaml
@@ -36,7 +36,7 @@ spec:
       group: gateway.kgateway.dev
       kind: Backend
 ---
-apiVersion: gateway.networking.k8s.io/v1beta1
+apiVersion: gateway.networking.k8s.io/v1
 kind: HTTPRoute
 metadata:
   name: gemini
